+++ NEWS.md
@@ -9,6 +9,8 @@
 * Fixed overall documentation and standardized package documentation.
 * Fixed test cases by adding seeds to reduce random noise, and fixed some style 
   issues.
+* `find_permutation_order` and `find_permutation_index` added to map the
+  permutation number to the actual permutation values and vice versa.
 * `cpp_profiler_tools.cpp` removed due to compilation issues on remote
   systems.